requirements.txt
openai>=1.0,<2.0
httpx>=0.25,<1.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
